+++ src/models/user.js
@@ -193,7 +193,7 @@ const userSchema = new mongoose.Schema(
   },
   {
     timestamps: true,
-  },
+  }
 );
 
 userSchema.index({ name: 'text', email: 'text' });